neuro_server/requirements.txt
Flask
gunicorn
numpy==1.26.4
pydub==0.25.1
pandas
pyannote.audio



